+++ spec/support/custom_object_changes_adapter.rb
@@ -21,4 +21,8 @@ class CustomObjectChangesAdapter
   def where_object_changes_from(klass, attributes)
     klass.where(attributes)
   end
+
+  def where_object_changes_to(klass, attributes)
+    klass.where(attributes)
+  end
 end