+++ CHANGELOG.md
@@ -7,22 +7,45 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Breaking Changes
 
-- In the PT rails engine, the `paper_trail` configuration has been
-  removed. This configuration object was deprecated in 10.2.0. Please review
-  docs section [2.d. Turning PaperTrail
-  Off](https://github.com/paper-trail-gem/paper_trail/#2d-turning-papertrail-off)
-  for alternatives.
+- None
 
 ### Added
 
 - `where_attribute_changes` queries for versions where the object's attribute
   changed to or from any values.
+
+### Fixed
+
+- None
+
+## 12.0.0 (2021-03-29)
+
+### Breaking Changes
+
+- Rails: The deprecated `config.paper_trail` configuration technique
+  has been removed. This configuration object was deprecated in 10.2.0. It only
+  had one key, `config.paper_trail.enabled`. Please review docs section [2.d.
+  Turning PaperTrail
+  Off](https://github.com/paper-trail-gem/paper_trail/#2d-turning-papertrail-off)
+  for alternatives.
+
+### Added
+
+- `where_object_changes_to` queries for versions where the object's attributes
+  changed to one set of known values from any other set of values.
 - `where_object_changes_from` queries for versions where the object's attributes
   changed from one set of known values to any other set of values.
 
 ### Fixed
 
+- [#1281](https://github.com/paper-trail-gem/paper_trail/pull/1281) Rails:
+  Instead of an `Engine`, PT now provides a `Railtie`, which is simpler.
 - Expand kwargs passed to `save_with_version` using double splat operator - Rails 6.1 compatibility
+- [#1287](https://github.com/paper-trail-gem/paper_trail/issues/1287) - Fix 'rails db:migrate' error when run against an app with mysql2 adapter
+
+- [#1285](https://github.com/paper-trail-gem/paper_trail/pull/1285) -
+  Touch callback does not create a new version for skipped or ignored attributes
+  for ActiveRecord 6.0 and higher
 
 ### Dependencies